refactor(questionnaires): extract SubmissionService from QuestionnaireService (#621)
questionnaire_service.py was 966 lines, within striking distance of the
1000-line gate. The single QuestionnaireService mixed two responsibilities:
authoring the questionnaire structure and ingesting/rendering submissions.

Split the submission half (build, submit, evaluate_submission, the submission
retrieval queries, and their private helpers) into a new SubmissionService in
questionnaires/service/submission_service.py. The shared prefetch moved to a
module-level load_questionnaire_with_relations() that both __init__s call, so
the loaded questionnaire is identical for both classes (no behavior change).

questionnaire_service.py drops to 480 lines; submission_service.py is 522.

Pure move/split — no logic changes. Call sites for submission flows (poll
detail build, public event submit/build, the three admin submission endpoints,
submit_event_questionnaire's param) now use SubmissionService; authoring sites
keep QuestionnaireService. Tests re-pointed accordingly.
