+++ src/events/tests/test_service/test_event_questionnaire_resubmission.py
@@ -23,7 +23,7 @@ from questionnaires.models import (
     QuestionnaireSubmission,
 )
 from questionnaires.schema import MultipleChoiceSubmissionSchema, QuestionnaireSubmissionSchema
-from questionnaires.service.questionnaire_service import QuestionnaireService
+from questionnaires.service.submission_service import SubmissionService
 
 pytestmark = pytest.mark.django_db
 
@@ -54,7 +54,7 @@ class TestAdmissionResubmissionValidation:
         self,
         eq_user: RevelUser,
         eq_event: Event,
-        eq_questionnaire_service: QuestionnaireService,
+        eq_questionnaire_service: SubmissionService,
         admission_org_questionnaire: OrganizationQuestionnaire,
         eq_mcq: MultipleChoiceQuestion,
         eq_option: MultipleChoiceOption,
@@ -85,7 +85,7 @@ class TestAdmissionResubmissionValidation:
         self,
         eq_user: RevelUser,
         eq_event: Event,
-        eq_questionnaire_service: QuestionnaireService,
+        eq_questionnaire_service: SubmissionService,
         admission_org_questionnaire: OrganizationQuestionnaire,
         eq_mcq: MultipleChoiceQuestion,
         eq_option: MultipleChoiceOption,
@@ -127,7 +127,7 @@ class TestAdmissionResubmissionValidation:
         self,
         eq_user: RevelUser,
         eq_event: Event,
-        eq_questionnaire_service: QuestionnaireService,
+        eq_questionnaire_service: SubmissionService,
         admission_org_questionnaire: OrganizationQuestionnaire,
         eq_mcq: MultipleChoiceQuestion,
         eq_option: MultipleChoiceOption,
@@ -173,7 +173,7 @@ class TestAdmissionResubmissionValidation:
         self,
         eq_user: RevelUser,
         eq_event: Event,
-        eq_questionnaire_service: QuestionnaireService,
+        eq_questionnaire_service: SubmissionService,
         admission_org_questionnaire: OrganizationQuestionnaire,
         eq_mcq: MultipleChoiceQuestion,
         eq_option: MultipleChoiceOption,
@@ -220,7 +220,7 @@ class TestAdmissionResubmissionValidation:
         eq_user: RevelUser,
         eq_event: Event,
         eq_questionnaire: Questionnaire,
-        eq_questionnaire_service: QuestionnaireService,
+        eq_questionnaire_service: SubmissionService,
         admission_org_questionnaire: OrganizationQuestionnaire,
         eq_mcq: MultipleChoiceQuestion,
         eq_option: MultipleChoiceOption,
@@ -271,7 +271,7 @@ class TestAdmissionResubmissionValidation:
         eq_user: RevelUser,
         eq_event: Event,
         eq_questionnaire: Questionnaire,
-        eq_questionnaire_service: QuestionnaireService,
+        eq_questionnaire_service: SubmissionService,
         admission_org_questionnaire: OrganizationQuestionnaire,
         eq_mcq: MultipleChoiceQuestion,
         eq_option: MultipleChoiceOption,
@@ -322,7 +322,7 @@ class TestAdmissionResubmissionValidation:
         eq_user: RevelUser,
         eq_event: Event,
         eq_questionnaire: Questionnaire,
-        eq_questionnaire_service: QuestionnaireService,
+        eq_questionnaire_service: SubmissionService,
         admission_org_questionnaire: OrganizationQuestionnaire,
         eq_mcq: MultipleChoiceQuestion,
         eq_option: MultipleChoiceOption,
@@ -379,7 +379,7 @@ class TestAdmissionResubmissionValidation:
         eq_user: RevelUser,
         eq_event: Event,
         eq_questionnaire: Questionnaire,
-        eq_questionnaire_service: QuestionnaireService,
+        eq_questionnaire_service: SubmissionService,
         admission_org_questionnaire: OrganizationQuestionnaire,
         eq_mcq: MultipleChoiceQuestion,
         eq_option: MultipleChoiceOption,
@@ -436,7 +436,7 @@ class TestAdmissionResubmissionValidation:
         eq_user: RevelUser,
         eq_event: Event,
         eq_questionnaire: Questionnaire,
-        eq_questionnaire_service: QuestionnaireService,
+        eq_questionnaire_service: SubmissionService,
         admission_org_questionnaire: OrganizationQuestionnaire,
         eq_mcq: MultipleChoiceQuestion,
         eq_option: MultipleChoiceOption,
@@ -492,7 +492,7 @@ class TestAdmissionResubmissionValidation:
         self,
         eq_user: RevelUser,
         eq_event: Event,
-        eq_questionnaire_service: QuestionnaireService,
+        eq_questionnaire_service: SubmissionService,
         admission_org_questionnaire: OrganizationQuestionnaire,
         eq_mcq: MultipleChoiceQuestion,
         eq_option: MultipleChoiceOption,
@@ -535,7 +535,7 @@ class TestAdmissionResubmissionValidation:
         eq_event: Event,
         eq_questionnaire: Questionnaire,
         eq_org: Organization,
-        eq_questionnaire_service: QuestionnaireService,
+        eq_questionnaire_service: SubmissionService,
         eq_mcq: MultipleChoiceQuestion,
         eq_option: MultipleChoiceOption,
     ) -> None:
