+++ src/questionnaires/service/__init__.py
@@ -4,9 +4,11 @@ from . import file_service
 from .duplication import collect_concrete_field_values, duplicate_questionnaire_content
 from .file_service import upload_questionnaire_file
 from .questionnaire_service import QuestionnaireService, get_questionnaire_schema
+from .submission_service import SubmissionService
 
 __all__ = [
     "QuestionnaireService",
+    "SubmissionService",
     "collect_concrete_field_values",
     "duplicate_questionnaire_content",
     "file_service",
